Workflow::Solr::Analyzer
* extract SolrBatch stuff to Workflow::Solr::Analyzer
* move Workflow::SolrDataSource to Workflow::SolrDataSource
* settings for milemarker & solr batch sizes
* continue renaming 'chunk size'
* fix nonworking phctl/workflow tests (from removing estimate stuff)
* SolrBatch always loads ht items in constructor -- avoids confusing
  forced lazy-load later, and we always need it in current use cases
* remove Workflows::SolrDataSource (moved to Workflows::Solr::DataSource)
* some other refactoring in overlap report data source
* use for estmate, cost report, overlap
