+++ config/settings.yml
@@ -20,3 +20,6 @@ solr_data_source:
   milemarker_batch_size: 50000
   solr_results_per_query: 5000
   ocns_per_solr_query: 500
+solr_analyzer:
+  milemarker_batch_size: 1000
+  solr_batch_size: 100
