+++ lib/sidekiq_jobs.rb
@@ -8,7 +8,6 @@ require "loader/shared_print_loader"
 require "reports/commitment_replacements"
 require "reports/cost_report"
 require "reports/dynamic"
-require "reports/frequency_table_from_solr"
 require "reports/holdings_by_date_report"
 require "reports/member_counts"
 require "reports/oclc_registration"
